+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hyperlane-xyz/radix-sdk
 
+## 25.3.0
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@25.3.0
+- @hyperlane-xyz/provider-sdk@1.3.4
+
 ## 25.2.0
 
 ### Patch Changes